requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.3
python-dotenv==1.1.0
jinja2==3.1.5
python-multipart==0.0.20
